fix(synth): update grounding-synthesis imports for module restructure
- conversation_synthesizer: pull describe_grounding_default from oumi.environments.utils (renamed from _helpers).
- test_conversation_synthesizer: pull GroundingConfig from oumi.core.configs.params.grounding_params (extracted from environment_params), pull DeterministicToolOutput / DeterministicTool from oumi.environments.deterministic_tool, and switch fixtures from ToolParams(deterministic_outputs=...) to DeterministicTool.
